+++ src/vulkan/ShaderBindingTable.hpp
@@ -19,7 +19,7 @@ struct ShaderBindingTable {
 	void create(const Device& device, const std::array<uint32_t, 4> entriesCount, VkPipeline pipeline) {
 		auto rayTracingPipelineProperties = device.getPhysicalDevice().getRaytracingPipelineProperties();
 
-		const size_t   totalEntries = entriesCount[0] + entriesCount[1] + entriesCount[2] + entriesCount[3];
+		const uint32_t totalEntries = entriesCount[0] + entriesCount[1] + entriesCount[2] + entriesCount[3];
 		const auto	   handle_size = rayTracingPipelineProperties.shaderGroupHandleSize;
 		const uint32_t handle_size_aligned = aligned_size(handle_size, rayTracingPipelineProperties.shaderGroupBaseAlignment);
 		const uint32_t regionSizes[4] = {
